make the fetch-source editbox flexible; make url a link
Manually resizing the editbox in response to the window being resized
causes the issue described in ticket #8287: the window grows
continuously. Instead, we give it a minimum height, and also make the
window resizable while snapping the size to the content size when we set
the URL to prevent the it or the close-button being cropped.